extract remaining mods, adapt gitignore, remove packages.config from solution
