+++ .gitignore
@@ -1,5 +1,5 @@
-**/bin
-**/obj
-**/.vs
+bin
+obj
+.vs
 *.user
 release/*
