+++ Mods/BackpackRenderer/Properties/launchSettings.json
@@ -1,0 +1,9 @@
+{
+  "profiles": {
+	"start-all-mods-in-sln": {
+      "commandName": "Executable",
+      "executablePath": "$(VINTAGE_STORY)/Vintagestory.exe",
+      "commandLineArgs": "--addModPath=\"$(SolutionDir)SourceMods\" \"$(SolutionDir)bin\" "
+    }
+  }
+}
