+++ Mods/PushPlayerJoinLeave/modinfo.json
@@ -1,0 +1,11 @@
+{
+  "type": "code",
+  "modid": "pushplayerjoinleave",
+  "name": "PushPlayerJoinLeave",
+  "author": "Unknown",
+  "description": ">>ADD YOUR DESCRIPTION HERE<<",
+  "version": "1.0.0",
+  "dependency": {
+    "game": "1.12.14"
+  }
+}
