--- assets/sampleassets/shapes/customshapes/plane.json
@@ -1,24 +1,0 @@
-{
-	"editor": {
-		"allAngles": false,
-		"singleTexture": false
-	},
-	"textureWidth": 16,
-	"textureHeight": 16,
-	"textures": {
-	},
-	"elements": [
-		{
-			"name": "Cube1",
-			"from": [ -2.0, 0.0, -2.0 ], 
-			"to": [ 18.0, 1.0, 18.0 ], 
-			"faces": {
-				"north": { "texture": "#null", "uv": [ 0.0, 0.0, 16.0, 1.0 ], "autoUv": false },
-				"east": { "texture": "#null", "uv": [ 0.0, 0.0, 16.0, 1.0 ], "autoUv": false },
-				"south": { "texture": "#null", "uv": [ 0.0, 0.0, 16.0, 1.0 ], "autoUv": false },
-				"west": { "texture": "#null", "uv": [ 0.0, 0.0, 16.0, 1.0 ], "autoUv": false },
-				"up": { "texture": "#null", "uv": [ 0.0, 0.0, 16.0, 16.0 ], "autoUv": false },
-				"down": { "texture": "#null", "uv": [ 0.0, 0.0, 16.0, 16.0 ], "autoUv": false }
-			}
-		}
-	]}
